Create Swidhifiedcycles
As the world changes do do I now we all know gas prices are getting in to our banks like never before. That's where I come in to bat. Using the wonderfully benefits of being in control of your gas spending is a great way to save more money than you realize. From my customers experiences an the advrage they have went from 80 dollars in gas to  6.00 within a two week period. Now that is a huge difference. Such savings has to have a catch well I am the type of person who is up front with exceptions.There is a catch a big one You must want it enough to obtain it! There is a few steps to take that I will walk you thru .1 step push your ego to the side(if you are too good to be seen on a bicycle turn back now )this is where you must make a choice within either keep driving your car all the time since you are so comfortable on it.Or take up the opertunity in your hands to ride to work l,save A ton of money. That was a no brained for me. Most of you are too proud of your cars to ride a bike so let be a realization .If you ride a bike built to your personality and needs. You can be the unique person you are without seeing another you on the road with money to spend. I build bicycles and not just any bicycles but custom bicycles with motors on them! I have countless combinations of sizes types and choices of electric,gas (both 2&4stroke) 2 and three week mods are available . The word custom takes a new step when you choose to get your hands on an exclusive swishified cycle. I work with each customer to choose the most extravagant bike just for them by our combined design. My bikes are like none on the plane to I use the best technology and personally hand machine the bearings to assure smooth operation that lasts like you can not find elsewhere. (Nothing rolls like a swishified). There is also special options that are exclusive to my bikes.i know this because these are my own inventions used to better the bikes praformance and comfortability combined with long lasting insurance to were and tear. I am confident in this confident enough to place a 2 year limited warranty on every bike.